[week8/refactor] 폴더 구조 변경 및 컴포넌트 정리
