+++ rene/Week8/Debouncing&Throttling/src/components/LPCard/LPCard.tsx
@@ -1,6 +1,7 @@
 import { useNavigate } from "react-router-dom";
 import type { LP } from "../../types/lp";
 import { getTimeAgo } from "../../utils/date";
+import { lpRecord as defaultThumbnail } from "../../assets";
 
 interface LPCardProps {
   lp: LP;
@@ -16,7 +17,7 @@ const LPCard = ({ lp }: LPCardProps) => {
       onClick={() => navigate(`/lps/${lp.id}`)}
     >
       <div className="absolute inset-0 overflow-hidden">
-        <img src={thumbnail} alt={title} className="w-full h-full object-cover" />
+        <img src={thumbnail || defaultThumbnail} alt={title || "thumbnail image"} className="w-full h-full object-cover" />
         <div className="absolute inset-0 bg-black/60 opacity-0 group-hover:opacity-100 transition-opacity flex flex-col justify-end p-3">
           <p className="text-white font-semibold text-sm truncate">{title}</p>
           <div className="flex items-center justify-between mt-1">
