+++ rene/Week8/Debouncing&Throttling/src/components/LPCard/LPGrid.tsx
@@ -3,17 +3,23 @@ import LPCard from "./LPCard";
 
 interface LPGridProps {
   lps: LP[];
+  gridClass?: string;
+  emptyMessage?: string;
 }
 
-const LPGrid = ({ lps }: LPGridProps) => {
+const LPGrid = ({
+  lps,
+  gridClass = "grid-cols-2 sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-5 xl:grid-cols-6",
+  emptyMessage = "등록된 LP가 없습니다.",
+}: LPGridProps) => {
   if (lps.length === 0) {
     return (
-      <div className="text-center py-20 text-zinc-500">등록된 LP가 없습니다.</div>
+      <div className="text-center py-20 text-zinc-500">{emptyMessage}</div>
     );
   }
 
   return (
-    <div className="grid grid-cols-2 sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-5 xl:grid-cols-6 gap-2">
+    <div className={`grid ${gridClass} gap-2`}>
       {lps.map((lp) => (
         <LPCard key={lp.id} lp={lp} />
       ))}
