+++ rene/Week8/Debouncing&Throttling/src/components/Layout/FloatingButton.tsx
@@ -1,4 +1,4 @@
-import plusIcon from "../assets/plus-icon.svg";
+import { plusIcon } from "../../assets";
 
 type FloatingButtonProps = {
   onClick?: () => void;
